+++ core/lib/sn/CmnInterface.d.ts
@@ -12,6 +12,7 @@ export declare type HArg = {
     scale_x?: number;
     scale_y?: number;
     visible?: boolean;
+    blendmode?: string;
     left?: number;
     top?: number;
     width?: number;
@@ -28,6 +29,7 @@ export declare type HArg = {
     time?: number;
     rule?: string;
     glsl?: string;
+    render?: boolean;
     pos?: string;
     text?: string;
     wait?: number;
@@ -65,7 +67,6 @@ export declare type HArg = {
     call?: boolean;
     global?: boolean;
     name?: string;
-    blendmode?: string;
     clear_local_event?: string;
     style?: string;
     style_hover?: string;
@@ -74,7 +75,7 @@ export declare type HArg = {
     b_height?: string;
     b_color?: string;
     b_alpha?: number;
-    b_alpha_isfixed?: string;
+    b_alpha_isfixed?: boolean;
     b_pic?: string;
     back_clear?: string;
     r_align?: string;